sbalance: Use a deferrable timer to avoid waking idle CPUs
SBalance is designed to poll to balance IRQs, but it shouldn't kick CPUs
out of idle to do so because idle CPUs clearly aren't processing
interrupts.

Open code a freezable wait that uses a deferrable timer in order to prevent
SBalance from waking up idle CPUs when there is little interrupt traffic.

Change-Id: I0269a251b0bbc6786e127c057a6e4d70d4ace335
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: angelomds42 <angelomds42@gmail.com>